Rewrote GUI for daq rates
