--- isstools/ui/ui_beamline_status.ui
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>Form</class>
- <widget class="QWidget" name="Form">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>336</width>
-    <height>134</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Form</string>
-  </property>
-  <widget class="QGroupBox" name="groupBox_shutters">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>0</y>
-     <width>321</width>
-     <height>131</height>
-    </rect>
-   </property>
-   <property name="sizePolicy">
-    <sizepolicy hsizetype="Expanding" vsizetype="Preferred">
-     <horstretch>0</horstretch>
-     <verstretch>0</verstretch>
-    </sizepolicy>
-   </property>
-   <property name="font">
-    <font>
-     <weight>50</weight>
-     <bold>false</bold>
-    </font>
-   </property>
-   <property name="title">
-    <string>Shutters</string>
-   </property>
-   <layout class="QHBoxLayout" name="horizontalLayout_shutters">
-    <property name="spacing">
-     <number>2</number>
-    </property>
-    <property name="sizeConstraint">
-     <enum>QLayout::SetMinimumSize</enum>
-    </property>
-    <property name="margin">
-     <number>2</number>
-    </property>
-    <item>
-     <layout class="QVBoxLayout" name="layout_shutters">
-      <property name="spacing">
-       <number>10</number>
-      </property>
-     </layout>
-    </item>
-   </layout>
-  </widget>
-  <zorder>groupBox_shutters</zorder>
-  <zorder></zorder>
- </widget>
- <resources/>
- <connections/>
-</ui>
