+++ isstools/ui/ui_xlive.ui
@@ -276,11 +276,11 @@
      <rect>
       <x>10</x>
       <y>10</y>
-      <width>1701</width>
+      <width>701</width>
       <height>131</height>
      </rect>
     </property>
-    <layout class="QHBoxLayout" name="layout_general_info">
+    <layout class="QHBoxLayout" name="layout_info_general">
      <property name="sizeConstraint">
       <enum>QLayout::SetNoConstraint</enum>
      </property>
@@ -295,7 +295,22 @@
       <height>131</height>
      </rect>
     </property>
-    <layout class="QHBoxLayout" name="layout_beamline_status"/>
+    <layout class="QHBoxLayout" name="layout_info_shutters"/>
+   </widget>
+   <widget class="QWidget" name="horizontalLayoutWidget_11">
+    <property name="geometry">
+     <rect>
+      <x>720</x>
+      <y>10</y>
+      <width>1091</width>
+      <height>131</height>
+     </rect>
+    </property>
+    <layout class="QHBoxLayout" name="layout_info_beamline">
+     <property name="sizeConstraint">
+      <enum>QLayout::SetNoConstraint</enum>
+     </property>
+    </layout>
    </widget>
   </widget>
   <widget class="QMenuBar" name="menubar">
