+++ isstools/widgets/widget_info_beamline.py
@@ -6,6 +6,7 @@ import urllib.request
 import numpy as np
 import os
 import re
+import time as ttime
 
 from isstools.dialogs import UpdateUserDialog, SetEnergy, GetEmailAddress
 from timeit import default_timer as timer
@@ -16,21 +17,22 @@ from issgoogletools.initialize import get_dropbox_service, get_gmail_service
 from issgoogletools.gmail import create_html_message, upload_draft, send_draft
 import uuid
 
-ui_path = pkg_resources.resource_filename('isstools', 'ui/ui_general_info.ui')
+ui_path = pkg_resources.resource_filename('isstools', 'ui/ui_info_beamline.ui')
 
 
 
 
-class UIGeneralInfo(*uic.loadUiType(ui_path)):
+class UIInfoBeamline(*uic.loadUiType(ui_path)):
+
     def __init__(self,
                  accelerator=None,
                  hhm = None,
                  shutters=None,
                  ic_amplifiers = None,
+                 apb = None,
                  RE = None,
                  db = None,
                  parent = None,
-
                  *args, **kwargs):
 
         super().__init__(*args, **kwargs)
@@ -41,30 +43,14 @@ class UIGeneralInfo(*uic.loadUiType(ui_path)):
         self.timer_update_time.timeout.connect(self.update_status)
         self.timer_update_time.start()
 
-        # self.timer_update_weather = QtCore.QTimer(self)
-        # self.timer_update_weather.singleShot(0, self.update_weather)
-        # self.timer_update_weather.setInterval(1000*60*5)
-        # self.timer_update_weather.timeout.connect(self.update_weather)
-        # self.timer_update_weather.start()
         self.hhm = hhm
         self.RE = RE
         self.db = db
         self.shutters = shutters
         self.ic_amplifiers = ic_amplifiers
         self.parent = parent
-
-        if self.RE is not None:
-            self.RE.is_aborted = False
-            self.timer_update_user_info = QtCore.QTimer()
-            self.timer_update_user_info.timeout.connect(self.update_user_info)
-            self.timer_update_user_info.start(60*1000)
-            self.timer_update_user_info.singleShot(0, self.update_user_info)
-            self.push_set_user_info.clicked.connect(self.set_user_info)
-            self.push_send_results.clicked.connect(self.send_results)
-
-        else:
-            self.push_update_user.setEnabled(False)
-
+        self.apb = apb
+        self.hhm= hhm
 
         # Initialize general settings
         self.accelerator = accelerator
@@ -80,25 +66,23 @@ class UIGeneralInfo(*uic.loadUiType(ui_path)):
         self.push_jog_pitch_neg.clicked.connect(self.tweak_pitch_neg)
         self.push_jog_pitch_pos.clicked.connect(self.tweak_pitch_pos)
 
-    def update_weather(self):
-        try:
-            current_weather = requests.get(
-                'http://api.openweathermap.org/data/2.5/weather?zip=11973&APPID=a3be6bc4eaf889b154327fadfd9d6532').json()
-            string_current_weather  = current_weather['weather'][0]['main'] + ' in Upton, NY,  it is {0:.0f} °F outside,\
-                humidity is {1:.0f}%'\
-                .format(((current_weather['main']['temp']-273)*1.8+32), current_weather['main']['humidity'])
-            icon_url = 'http://openweathermap.org/img/w/' + current_weather['weather'][0]['icon'] + '.png'
-            image = QtGui.QImage()
-            image.loadFromData(urllib.request.urlopen(icon_url).read())
-            self.label_current_weather_icon.setPixmap(QtGui.QPixmap(image))
-        except:
-            string_current_weather = 'Weather information not available'
-        self.label_current_weather.setText(string_current_weather)
+
+
+        daq_rate = self.apb.acq_rate.get()
+        self.spinBox_daq_rate.setValue(daq_rate)
+        self.spinBox_daq_rate.valueChanged.connect(self.update_daq_rate)
+
+        enc_rate_in_points = hhm.enc.filter_dt.get()
+        enc_rate = 1/(89600*10*1e-9)/1e3
+        self.spinBox_enc_rate.setValue(enc_rate)
+        self.spinBox_enc_rate.valueChanged.connect(self.update_enc_rate)
+
+
+
 
 
     def update_status(self):
-        self.label_current_time.setText(
-            'Today is {0}'.format(QtCore.QDateTime.currentDateTime().toString('MMMM d, yyyy, h:mm:ss ap')))
+
         energy = self.hhm.energy.read()['hhm_energy']['value']
         self.label_energy.setText('Energy is {:.1f} eV'.format(energy))
         if ((self.hhm.fb_status.get()==1) and
@@ -134,8 +118,6 @@ class UIGeneralInfo(*uic.loadUiType(ui_path)):
             self.label_RE_status_indicator.setStyleSheet('background-color: rgb(255,0,0)')
 
 
-
-
     def update_beam_current(self, **kwargs):
         self.label_beam_current.setText('Beam current is {:.1f} mA'.format(kwargs['value']))
 
@@ -169,78 +151,6 @@ class UIGeneralInfo(*uic.loadUiType(ui_path)):
             self.label_accelerator_status.setStyleSheet('color: rgb(19,139,67)')
             self.label_accelerator_status_indicator.setStyleSheet('background-color: rgb(0,177,0)')
 
-    def update_user_info(self):
-        self.label_user_info.setText('{} is running  under Proposal {}/SAF {} '.
-                                     format(self.RE.md['PI'], self.RE.md['PROPOSAL'], self.RE.md['SAF']))
-        self.cycle = ['', 'Spring', 'Summer', 'Fall']
-        self.label_current_cycle.setText(
-            'It is {} {} NSLS Cycle'.format(self.RE.md['year'], self.cycle[int(self.RE.md['cycle'])]))
-
-    def set_user_info(self):
-        dlg = UpdateUserDialog.UpdateUserDialog(self.RE.md['year'], self.RE.md['cycle'], self.RE.md['PROPOSAL'],
-                                                self.RE.md['SAF'], self.RE.md['PI'], parent=self)
-        if dlg.exec_():
-            start = timer()
-            self.RE.md['year'], self.RE.md['cycle'], self.RE.md['PROPOSAL'], self.RE.md['SAF'], self.RE.md[
-                'PI'] = dlg.getValues()
-            stop1 = timer()
-            self.update_user_info()
-
-
-
-
-    def send_results(self):
-
-        dlg = GetEmailAddress.GetEmailAddress('', parent=self)
-        if dlg.exec_():
-            email_address = dlg.getValue()
-            regex = '^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$'
-            if re.search(regex, email_address):
-                 #print(f'email {email_address}')
-                pass
-            else:
-                message_box('Error', 'Invaild email')
-                return 0
-
-
-        year=self.RE.md['year']
-        cycle=self.RE.md['cycle']
-        proposal = self.RE.md['PROPOSAL']
-        PI = self.RE.md['PI']
-        working_directory = f'/nsls2/xf08id/users/{year}/{cycle}/{proposal}'
-        zip_file = f'{working_directory}/{proposal}.zip'
-
-        id = str(uuid.uuid4())[0:5]
-
-        zip_id_file = f'{proposal}-{id}.zip'
-
-        if os.path.exists(zip_file):
-            os.remove(zip_file)
-
-        os.system(f'zip {zip_file} {working_directory}/*.dat ')
-        dropbox_service =  get_dropbox_service()
-        with open(zip_file,"rb") as f:
-            file_id = dropbox_service.files_upload(f.read(),f'/{year}/{cycle}/{zip_id_file}')
-
-
-        file_link = dropbox_service.sharing_create_shared_link(f'/{year}/{cycle}/{zip_id_file}')
-        link_url =  file_link.url
-        print('Upload succesful')
-
-        gmail_service = get_gmail_service()
-        message = create_html_message(
-            'staff08id@gmail.com',
-            email_address,
-            f'ISS beamline results Proposal {proposal}',
-            f' <p> Dear {PI},</p> <p>You can download the result of your experiment at ISS under proposal {proposal} here,</p> <p> {link_url} </p> <p> Sincerely, </p> <p> ISS Staff </p>'
-            )
-
-        draft = upload_draft(gmail_service, message)
-        sent = send_draft(gmail_service, draft)
-        print('Email sent')
-
-
-
 
     def set_i0_gain(self):
         self.ic_amplifiers['i0_amp'].set_gain(int(self.comboBox_set_i0_gain.currentText()),0)
@@ -268,9 +178,6 @@ class UIGeneralInfo(*uic.loadUiType(ui_path)):
             except Exception as exc:
                 message_box('Incorrect energy','Energy should be within 4700-32000 eV range')
 
-
-
-
     def tweak_pitch_pos(self):
         self.parent.widget_beamline_setup.pushEnableHHMFeedback.setChecked(False)
         pitch = self.hhm.pitch.read()['hhm_pitch']['value']
@@ -280,3 +187,20 @@ class UIGeneralInfo(*uic.loadUiType(ui_path)):
         self.parent.widget_beamline_setup.pushEnableHHMFeedback.setChecked(False)
         pitch = self.hhm.pitch.read()['hhm_pitch']['value']
         self.RE(bps.mv(self.hhm.pitch, pitch-0.025))
+
+    def update_daq_rate(self):
+        daq_rate = self.spinBox_daq_rate.value()
+        # 374.94 is the nominal RF frequency
+        divider = int(374.94/daq_rate)
+        self.RE(bps.abs_set(self.apb.divide, divider, wait=True))
+
+    def update_enc_rate(self):
+        enc_rate = self.spinBox_enc_rate.value()
+        rate_in_points = (1/(enc_rate*1e3))*1e9/10
+        print(rate_in_points)
+        #self.RE(bps.abs_set(self.hhm.enc.filter_dt, rate_in_points, wait=True))
+
+
+
+
+
