+++ isstools/widgets/widget_run.py
@@ -55,9 +55,8 @@ class UIRun(*uic.loadUiType(ui_path)):
                  db,
                  hhm,
                  shutter_dictionary,
-                 adc_list,
-                 enc_list,
-                 xia,
+
+
                  apb,
                  parent_gui,
 
@@ -74,9 +73,8 @@ class UIRun(*uic.loadUiType(ui_path)):
         self.db = db
         self.hhm=hhm,
         self.shutter_dictionary = shutter_dictionary
-        self.adc_list = adc_list
-        self.enc_list = enc_list
-        self.xia = xia
+
+
         self.apb = apb
         self.parent_gui = parent_gui
         self.comboBox_scan_type.addItems(self.plan_funcs_names)
@@ -201,8 +199,6 @@ class UIRun(*uic.loadUiType(ui_path)):
             if self.rr_token is not None:
                 self.RE.unsubscribe(self.rr_token)
 
-
-
         else:
             message_box('Error', 'Please provide the name for the scan')
 
@@ -219,15 +215,6 @@ class UIRun(*uic.loadUiType(ui_path)):
         for i in range(len(self.parameter_values)):
             self.gridLayout_parameters.addWidget(self.parameter_values[i], i, 0, QtCore.Qt.AlignTop)
             self.gridLayout_parameters.addWidget(self.parameter_descriptions[i], i, 1, QtCore.Qt.AlignTop)
-
-    def setAnalogSampTime(self, text):
-        self.analog_samp_time = text
-
-    def setEncSampTime(self, text):
-        self.enc_samp_time = text
-
-    def setXiaSampTime(self, text):
-        self.xia_samp_time = text
 
     def draw_interpolated_data(self, df):
         update_figure([self.figure.ax2, self.figure.ax1, self.figure.ax3], self.toolbar, self.canvas)
